+++ langgraph_agent_toolkit/schema/__init__.py
@@ -1,8 +1,12 @@
 from langgraph_agent_toolkit.schema.schema import (
+    AddMessagesInput,
+    AddMessagesResponse,
     AgentInfo,
     ChatHistory,
     ChatHistoryInput,
     ChatMessage,
+    ClearHistoryInput,
+    ClearHistoryResponse,
     Feedback,
     FeedbackResponse,
     HealthCheck,
@@ -13,9 +17,13 @@ from langgraph_agent_toolkit.schema.schema import (
 
 
 __all__ = [
+    "AddMessagesInput",
+    "AddMessagesResponse",
     "AgentInfo",
     "UserInput",
     "ChatMessage",
+    "ClearHistoryInput",
+    "ClearHistoryResponse",
     "ServiceMetadata",
     "StreamInput",
     "Feedback",