+++ CHANGELOG.md
@@ -6,6 +6,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Updated
+
+- Endpoints
+
+### Added
+
+- Endpoint to clear history
+- Add new message to the history
+
+### Fixed
+
+- Minor fixes and refactoring
+
+## [0.3.1]
+
+### Added
+
+- Ability to pass parameters to service runner
+- Argument to select service runner
+
+### Updated
+
+- Service Dockerfile
+
 ## [0.3.0]
 
 ### Added